# app/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para o Tech Challenge 2 da FIAP - SOAT",
        "title": "Lanchonete API - Tech Challenge 2",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/acompanhamento": {
            "post": {
                "description": "Cria um acompanhamento",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "acompanhamento"
                ],
                "summary": "Cria um acompanhamento",
                "parameters": [
                    {
                        "description": "Acompanhamento",
                        "name": "acompanhamento",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entities.AcompanhamentoPedido"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/acompanhamento/show": {
            "get": {
                "description": "Busca um acompanhamento",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "acompanhamento"
                ],
                "summary": "Busca um acompanhamento",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do acompanhamento",
                        "name": "ID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/presenters.AcompanhamentoDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/acompanhamento/{IDAcompanhamento}/{IDPedido}": {
            "post": {
                "description": "Adiciona um pedido existente ao acompanhamento de pedidos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "acompanhamento"
                ],
                "summary": "Adiciona um pedido ao acompanhamento",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do acompanhamento",
                        "name": "IDAcompanhamento",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID do pedido",
                        "name": "IDPedido",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Pedido ou acompanhamento não encontrado",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Erro interno",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/acompanhamento/{IDAcompanhamento}/{IDPedido}/{status}": {
            "put": {
                "description": "Atualiza o status de um pedido",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "acompanhamento"
                ],
                "summary": "Atualiza o status de um pedido",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do acompanhamento",
                        "name": "IDAcompanhamento",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID do pedido",
                        "name": "IDPedido",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Novo status",
                        "name": "status",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/acompanhamento/{ID}": {
            "get": {
                "description": "Busca um pedido",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "acompanhamento"
                ],
                "summary": "Busca um pedido",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do pedido",
                        "name": "ID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.AcompanhamentoPedido"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/cliente": {
            "post": {
                "description": "Cria um cliente",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cliente"
                ],
                "summary": "Cria um cliente",
                "parameters": [
                    {
                        "description": "Cliente",
                        "name": "cliente",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entities.Cliente"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/cliente/{CPF}": {
            "get": {
                "description": "Busca um cliente",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cliente"
                ],
                "summary": "Busca um cliente",
                "parameters": [
                    {
                        "type": "string",
                        "description": "CPF do cliente",
                        "name": "CPF",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.Cliente"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/pagamento": {
            "post": {
                "description": "Envia o pagamento para o webhook",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pagamento"
                ],
                "summary": "Envia o pagamento para o webhook",
                "parameters": [
                    {
                        "description": "Pagamento",
                        "name": "pagamento",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entities.Pagamento"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/pagamento/confirmar": {
            "post": {
                "description": "Confirma o pagamento",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pagamento"
                ],
                "summary": "Confirma o pagamento",
                "parameters": [
                    {
                        "description": "Pagamento",
                        "name": "pagamento",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entities.Pagamento"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "entities.AcompanhamentoPedido": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "pedidos": {
                    "$ref": "#/definitions/entities.FilaPedidos"
                },
                "tempoEstimado": {
                    "$ref": "#/definitions/time.Duration"
                },
                "ultimaAtualizacao": {
                    "type": "string"
                }
            }
        },
        "entities.CatProduto": {
            "type": "string",
            "enum": [
                "Lanche",
                "Acompanhamento",
                "Bebida",
                "Sobremesa"
            ],
            "x-enum-varnames": [
                "Lanche",
                "Acompanhamento",
                "Bebida",
                "Sobremesa"
            ]
        },
        "entities.Cliente": {
            "type": "object",
            "properties": {
                "cpf": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "nome": {
                    "type": "string"
                }
            }
        },
        "entities.FilaPedidos": {
            "type": "object",
            "properties": {
                "pedidos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.Pedido"
                    }
                }
            }
        },
        "entities.Pagamento": {
            "type": "object",
            "properties": {
                "dataCriacao": {
                    "type": "string"
                },
                "idPagamento": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "valor": {
                    "type": "string"
                }
            }
        },
        "entities.Pedido": {
            "type": "object",
            "properties": {
                "cliente": {
                    "$ref": "#/definitions/entities.Cliente"
                },
                "identificacao": {
                    "type": "string"
                },
                "personalizacao": {
                    "type": "string"
                },
                "produtos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.Produto"
                    }
                },
                "status": {
                    "$ref": "#/definitions/entities.StatusPedido"
                },
                "statusPagamento": {
                    "type": "string"
                },
                "timeStamp": {
                    "type": "string",
                    "format": "date-time"
                },
                "total": {
                    "type": "number"
                },
                "ultimaAtualizacao": {
                    "type": "string",
                    "format": "date-time"
                }
            }
        },
        "entities.Produto": {
            "type": "object",
            "properties": {
                "categoria": {
                    "$ref": "#/definitions/entities.CatProduto"
                },
                "descricao": {
                    "type": "string"
                },
                "identificacao": {
                    "type": "string"
                },
                "nome": {
                    "type": "string"
                },
                "preco": {
                    "type": "number"
                }
            }
        },
        "entities.StatusPedido": {
            "type": "string",
            "enum": [
                "Pendente",
                "Recebido",
                "Em preparação",
                "Pronto",
                "Finalizado"
            ],
            "x-enum-varnames": [
                "Pendente",
                "Recebido",
                "EmPreparacao",
                "Pronto",
                "Finalizado"
            ]
        },
        "presenters.AcompanhamentoDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "pedidos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/presenters.PedidoDTO"
                    }
                },
                "tempoEstimado": {
                    "description": "in minutes",
                    "type": "integer"
                }
            }
        },
        "presenters.PedidoDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "identificacao": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/entities.StatusPedido"
                },
                "tempoEstimado": {
                    "$ref": "#/definitions/time.Duration"
                }
            }
        },
        "response.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "response.SuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "time.Duration": {
            "type": "integer",
            "enum": [
                -9223372036854775808,
                9223372036854775807,
                1,
                1000,
                1000000,
                1000000000,
                60000000000,
                3600000000000
            ],
            "x-enum-varnames": [
                "minDuration",
                "maxDuration",
                "Nanosecond",
                "Microsecond",
                "Millisecond",
                "Second",
                "Minute",
                "Hour"
            ]
        }
    }
}